feat: implement store models, serializers, views, and frontend checkout/profile pages with gift card support

+++ backend/store/models.py
@@ -74,7 +74,7 @@ class Customer(models.Model):
     birth_date = models.DateField(null=True, blank=True)
     membership = models.CharField(
         max_length=1, choices=MEMBERSHIP_CHOICES, default=MEMBERSHIP_BRONZE)
-    vibe_coin = models.PositiveIntegerField(default=0)
+    vibe_coin = models.DecimalField(max_digits=10, decimal_places=2, default=0.00)
     user = models.OneToOneField(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
 
     def __str__(self):
